backend: more consolidation
