+++ AGENTS.md
@@ -64,7 +64,7 @@ The below rules are derived from `docs/adr/`. When applying the concepts, you sh
   - React components and feature pages: PascalCase files like `TextInput.tsx`
   - Utilities: camelCase files like `timeAgo.ts`
   - Route files: TanStack Router flat-file naming like `_site.games.$gameId.tsx` or `_game.games.$gameId.play.tsx`
-- Backend code relies on `erasableSyntaxOnly`, so do not use TypeScript `enum` in the backend. Use `as const` objects plus derived union types, following `backend/src/lib/gameResources.ts`.
+- Backend code relies on `erasableSyntaxOnly`, so do not use TypeScript `enum` in the backend. Use `as const` objects and the `Enumify` type helper to create the enum type.
 - Do not create shared utility/helper files or folders. Create dedicated files with actual names instead of generic files.
 - When importing a value and a type of the same name from the same package (e.g `Range` or `Result`), just import the value. Do not import the type to rename it.
