+++ backend/src/api/createApi.ts
@@ -15,9 +15,9 @@ import { LobbiesController } from "#api/lobbies/lobbies.controller.ts"
 import { type LobbiesRepository } from "#api/lobbies/lobbies.repository.ts"
 import { createLobbiesRouter } from "#api/lobbies/lobbies.router.ts"
 import { StarSystemsController } from "#api/star-systems/starSystems.controller.ts"
+import type { StarSystemsRepository } from "#api/star-systems/starSystems.repository.ts"
 import { createStarSystemsRouter } from "#api/star-systems/starSystems.router.ts"
 import type { CreateTransaction } from "#lib/db/createDb.ts"
-import type { StarSystemsRepository } from "#lib/db/star-systems/starSystems.repository.ts"
 import { requestLoggerMiddleware } from "./requestLoggerMiddleware.ts"
 import { createTrpc, createTrpcContext } from "./trpc.ts"
