+++ backend/src/api/entry.api.ts
@@ -6,10 +6,10 @@ import { AuthService } from "#api/accounts/auth.service.ts"
 import { GameplayRepository } from "#api/gameplay/gameplay.repository.ts"
 import { ListingsRepository } from "#api/listings/listings.repository.ts"
 import { LobbiesRepository } from "#api/lobbies/lobbies.repository.ts"
+import { StarSystemsRepository } from "#api/star-systems/starSystems.repository.ts"
 import { configureLogger } from "#lib/configureLogger.ts"
 import type { Database } from "#lib/db/createDb.ts"
 import { createDb } from "#lib/db/createDb.ts"
-import { StarSystemsRepository } from "#lib/db/star-systems/starSystems.repository.ts"
 import { envSchema, parseEnv } from "#lib/parseEnv.ts"
 import { startTickProcessing } from "#tick-processing/entry.tick-processing.ts"
 import { createApi } from "./createApi.ts"
