+++ backend/src/api/lobbies/lobbies.router.test.ts
@@ -3,7 +3,7 @@ import { createNewAccountModelStub } from "#api/accounts/NewAccountModel.stub.ts
 import { createApiStub } from "#api/createApi.stub.ts"
 import { type CreateLobbyDto, type LobbyPlayerDto } from "#api/lobbies/lobbies.controller.ts"
 import { GameStatus } from "#api/shared/GameStatus.ts"
-import { createDefaultStarSystemGenerationSettings } from "#lib/star-systems/createDefaultStarSystemGenerationSettings.ts"
+import { createDefaultStarSystemGenerationSettings } from "#api/star-systems/createDefaultStarSystemGenerationSettings.ts"
 import { extractSuccess } from "#tests/extractSuccess.ts"
 import { TrpcClient } from "#tests/TrpcClient.ts"
