+++ backend/src/api/star-systems/StarSystemGenerationSettings.stub.ts
